requirements.txt

streamlit>=1.28.0
pandas>=2.0.0
openai>=1.0.0
anthropic>=0.7.0
python-docx>=0.8.11
markdown>=3.4.0
PyYAML>=6.0
tiktoken>=0.7.0
pytest>=7.0.0
tinydb>=4.8.0
